feat(part4): complete bloglist expansion exercises 4.15-4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,17 @@
 const blogsRouter = require("express").Router();
 
 const Blog = require("../models/blog");
+const User = require("../models/user");
+
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
+  const { user } = request;
   const { title, author, url, likes = 0 } = request.body;
 
   if (!title || !author || !url) {
@@ -16,9 +20,19 @@ blogsRouter.post("/", async (request, response) => {
     });
   }
 
-  const blog = new Blog({ title, author, url, likes });
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user._id,
+  });
 
   const savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
@@ -40,7 +54,17 @@ blogsRouter.put("/:id", async (request, response) => {
   response.json(updatedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).end();
+  }
+
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(403).json({ error: "forbidden" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
